# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateUserRequest:
    properties:
      full_name:
        description: Full name untuk display (optional, default = username jika kosong)
        example: John Sales Manager
        type: string
      is_active:
        description: Pointer untuk default true jika tidak dikirim
        example: true
        type: boolean
      password:
        description: Password plain text (akan di-hash dengan bcrypt)
        example: password123456
        type: string
      role:
        description: 'Role: ''super_admin'', ''admin_sales'', atau ''admin_ops'''
        example: admin_sales
        type: string
      username:
        description: Username (wajib, unique)
        example: john.sales
        type: string
    required:
    - password
    - role
    - username
    type: object
...
  model.UpdateProfileRequest:
    properties:
      full_name:
        description: Update display name (optional)
        example: Bambang Suryanto
        type: string
      password:
        description: Update password (optional, minimal 6 karakter jika diisi)
        example: newpassword123456
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jalurberlian.id
    name: API Support
    url: http://www.jalurberlian.id/support
  description: Real-time fleet management and order tracking system for PT. Jalur
    Berlian Makassar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Jalur Berlian Fleet Management API
  version: "1.0"
paths:
...
  /admin/profile:
    patch:
      consumes:
      - application/json
      description: Update your own profile information (full_name and/or password).
        All authenticated admins can access. At least one field must be provided.
        Password must be at least 6 characters. User_id is extracted from JWT token.
      parameters:
      - description: 'Profile update data: full_name (optional), password (optional,
          min 6 chars). At least one required.'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully with message and updated user
            object (id, username, full_name, role, is_active, created_at)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad request: at least one field (full_name or password) must
            be provided, or password must be at least 6 characters'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found - user not found in database
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable entity - failed to update profile
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update own profile (all admin roles)
      tags:
      - User Management
...
  /admin/users:
    get:
      description: Get list of all users with their username, role, and active status
        (password_hash excluded). Only super_admin can access. Returns message, count,
        and data array.
      produces:
      - application/json
      responses:
        "200":
          description: Success response with message, count, and data array of users
            (each user has id, username, full_name, role, is_active, created_at)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - super_admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List all users (super_admin only)
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a new internal admin user with username, password, and role.
        Only super_admin can access. Full_name defaults to username if not provided.
        Password must be at least 6 characters.
      parameters:
      - description: 'User creation data: username (required), password (required,
          min 6 chars), full_name (optional), role (required: super_admin, admin_sales,
          or admin_ops), is_active (optional, default true)'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully with user object containing id, username,
            full_name, role, is_active, created_at
          schema:
            type: object
        "400":
          description: 'Invalid input: username required, password required, password
            must be at least 6 characters, role must be one of: super_admin, admin_sales,
            admin_ops'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - super_admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - username already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable entity - failed to create user
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create new admin user (super_admin only)
      tags:
      - User Management
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"